+++ src/main/kotlin/dev/rochajg/domain/gateway/transaction/TransactionGateway.kt
@@ -5,5 +5,5 @@ import dev.rochajg.domain.entity.transaction.Transaction
 interface TransactionGateway {
     fun create(transaction: Transaction)
 
-    fun getByUser(userId: Int): List<Transaction>
+    fun getByUser(userId: Long): List<Transaction>
 }
