+++ src/main/kotlin/dev/rochajg/domain/gateway/user/UserGateway.kt
@@ -3,7 +3,7 @@ package dev.rochajg.domain.gateway.user
 import dev.rochajg.domain.entity.user.User
 
 interface UserGateway {
-    fun get(id: Int?): User?
+    fun get(id: Long?): User?
 
     fun update(user: User)
 }
